requirements.txt
pytelegrambotapi==4.19.1
requests==2.31.0
schedule==1.2.1
qrcode==7.4.2
pillow==11.3.0
pydub==0.25.1
speechrecognition==3.10.0
python-dotenv==1.0.0